+++ plugins/admin/classes/plugin/Controllers/Login/LoginController.php
@@ -3,7 +3,7 @@
 /**
  * @package    Grav\Plugin\Admin
  *
- * @copyright  Copyright (c) 2015 - 2023 Trilby Media, LLC. All rights reserved.
+ * @copyright  Copyright (c) 2015 - 2024 Trilby Media, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
 
@@ -469,7 +469,15 @@ class LoginController extends AdminController
         $fullname = $user->fullname ?: $username;
         $author = $config->get('site.author.name', '');
         $sitename = $config->get('site.title', 'Website');
-        $reset_link = $this->getAbsoluteAdminUrl("/reset/u/{$username}/{$token}");
+        $reset_route = "/reset/u/{$username}/{$token}";
+
+        $site_host = $config->get('plugins.login.site_host');
+        if (!empty($site_host)) {
+            $admin = $this->getAdmin();
+            $reset_link = rtrim($site_host, '/') . '/' . trim($admin->base, '/') . '/' . ltrim($reset_route, '/');
+        } else {
+            $reset_link = $this->getAbsoluteAdminUrl($reset_route);
+        }
 
         // For testing only!
         //Admin::DEBUG && Admin::addDebugMessage(sprintf('Reset link: %s', $reset_link));