+++ plugins/form/form.php
@@ -518,7 +518,7 @@ class FormPlugin extends Plugin
                 $captcha_value = trim($form->value('basic-captcha'));
                 if (!$captcha->validateCaptcha($captcha_value)) {
                     $message = $params['message'] ?? $this->grav['language']->translate('PLUGIN_FORM.ERROR_BASIC_CAPTCHA');
-
+                    $form->setData('basic-captcha', '');
                     $this->grav->fireEvent('onFormValidationError', new Event([
                         'form' => $form,
                         'message' => $message
@@ -665,6 +665,11 @@ class FormPlugin extends Plugin
                     }
 
                     $filename = $prefix . $this->udate($format, $raw_format) . $postfix . $ext;
+                }
+
+                // Handle bad filenames.
+                if (!Utils::checkFilename($filename)) {
+                    throw new RuntimeException(sprintf('Form save: File with extension not allowed: %s', $filename));
                 }
 
                 /** @var Twig $twig */
@@ -1130,6 +1135,10 @@ class FormPlugin extends Plugin
                 return false;
             }
 
+            if (isset($form->xhr_submit) && $form->xhr_submit) {
+                $form->set('template', $form->template ?? 'form-xhr');
+            }
+
             // Set page template if passed by form
             if (isset($form->template)) {
                 $this->grav['page']->template($form->template);
@@ -1290,7 +1299,10 @@ class FormPlugin extends Plugin
     {
         /** @var \Grav\Common\Cache $cache */
         $cache = $this->grav['cache'];
-        $cache_id = $cache->getKey() . '-form-plugin';
+        /** @var Pages $pages */
+        $pages= $this->grav['pages'];
+//        $cache_id = $cache->getKey() . '-form-plugin';
+        $cache_id = $pages->getPagesCacheId() . '-form-plugin';
         return $cache_id;
     }