+++ plugins/table-importer/CHANGELOG.md
@@ -1,4 +1,21 @@
-#v2.2.1
+# v3.0.0
+## 09/07/2024
+
+1. [](#new)
+    * Added table footer param and logic
+1. [](#improved)
+    * Rewrote table HTML output logic to use DOMElements rather than string concatenation
+    * Cleaned up param parsing code
+
+# v2.3.0
+## 04/23/2024
+
+1. [](#new)
+    * Updated League CSV to version 9.8 (drops support for PHP < v7.4) 
+    * Updated CSV logic for new lib methods
+    * Updated Readme
+
+# v2.2.1
 ## 11/12/2019
 
 1. [](#bugfix)
@@ -9,7 +26,6 @@
 
 1. [](#new)
     * You can now add custom `id` tags to the tables (thanks to @mwender).
-
 1. [](#improved)
     * Improved error reporting when things go wrong.